Replaced Resources with Fluent API
